feat/repository: repository 파일 추가

+++ src/main/kotlin/com/jordyma/blink/folder/repository/FolderRepository.kt
@@ -3,5 +3,5 @@ package com.jordyma.blink.folder.repository
 import com.jordyma.blink.folder.entity.Folder
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface FolderRepository : JpaRepository<Folder, Int> {
+interface FolderRepository : JpaRepository<Folder, Long> {
 }
